@@ -27,7 +27,7 @@ filterwarnings = [
 
 [tool.ruff]
 line-length = 88
-target-version = "py39"
+target-version = "py310"
 extend-exclude = ["docs/*", "local_ignore/*"]
 
 [tool.ruff.lint]
