@@ -6,7 +6,6 @@ import inspect
 import logging
 import tempfile
 from pathlib import Path
-from typing import Optional
 
 import pandas as pd
 
@@ -33,7 +32,7 @@ class RunResult:
         operation: str,
         operation_descr: str,
         secs_taken: float,
-        run_details: Optional[dict] = None,
+        run_details: dict | None = None,
     ):
         """Constructor for a RunResult.
 
@@ -60,8 +59,8 @@ class RunResult:
 
 
 def run_benchmarks(
-    modules_to_run: Optional[list[str]] = None,
-    functions_to_run: Optional[list[str]] = None,
+    modules_to_run: list[str] | None = None,
+    functions_to_run: list[str] | None = None,
 ):
     """Run all benchmarks specified.
 
