@@ -5,7 +5,7 @@ import math
 import shutil
 import tempfile
 from pathlib import Path
-from typing import Literal, Optional
+from typing import Literal
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -109,8 +109,8 @@ def save_chart(
     df: pd.DataFrame,
     title: str,
     output_path: Path,
-    yscale: Optional[Literal["linear", "log", "symlog", "logit"]] = None,
-    y_value_formatter: Optional[str] = None,
+    yscale: Literal["linear", "log", "symlog", "logit"] | None = None,
+    y_value_formatter: str | None = None,
     print_labels_on_points: bool = False,
     size: tuple[float, float] = (8, 4),
     plot_kind: Literal[
@@ -126,8 +126,8 @@ def save_chart(
         "scatter",
         "hexbin",
     ] = "line",
-    gridlines: Optional[Literal["both", "x", "y"]] = None,
-    linestyle: Optional[str] = None,
+    gridlines: Literal["both", "x", "y"] | None = None,
+    linestyle: str | None = None,
 ):
     """Render and save a chart.
 
